fix issue wit re-mounting ClusterNodeAddPage
This is a temporary fix. It should be reworked in the near future.

The proble is with Route#component. According react-router
docummentation:

When you use component the router uses React.createElement to create a
new React element from the given component. That means if you provide an
inline function to the component prop, you would create a new component
every render. This results in the existing component unmounting and the
new component mounting instead of just updating the existing component.

It turned out that the add node form was reset everytime the cluster
status arrived.