fix an issue with expandable button display
This fixes the problem where an icon indicating summary status had been
placed under number indicating items count instead of placing icon next
to number - in dashboard, in expandable cell with count of nodes,
resources, etc...

Value `inline-block` has been applied to a css property `display` of the
expandable button. Actually value `flex` has been applied (because of
`.pf-l-split`) but `inline-block` (because of `.pf-c-button`) won. This
fix applies `flex` explicitly and so it fix the problem. However, it
should be somehow cleaned up in the future.