make cluster status types more reality suitable
When cluster status gaining mechanism is not able to connect to node it
returns some default status for the node. It looks like the node status
is correctly gained. This patch explicitly separate situation when data
was gained correctly from situation when data could not be obtained.

So far it looked like the resource status can be "unknown". But resource
status cannot be "unknown" actually.