fix: remove svgo (css-minimizer-webpack-plugin)
Compiled version of svgo contains forbidden unicode characters. It is
probably not real security issue, however, it is possible to avoid this
library (+ 54 another subdependencies).

Instead of direct dependency css-minimizer-webpack-plugin only csso was
added. The package csso has been already present here since it is a deep
subdependency of svgo which is dependency of
css-minimizer-webpack-plugin.